requirements.txt
selenium>=4.16.0,<5.0.0
undetected-chromedriver>=3.5.0
pyautogui>=0.9.53
python-dotenv>=1.0.0
